% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acre-package.R
\docType{package}
\name{acre-package}
\alias{acre}
\alias{acre-package}
\title{\code{acre}: Acoustic spatial capture-recapture models}
\description{
The \code{acre} package provides functions to fit spatial
capture-recapture models, with a particular focus on analysing data
from acoustic surveys.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/b-steve/acre}
  \item \url{https://b-steve.github.io/acre/}
}

}
\author{
\strong{Maintainer}: Ben Stevenson \email{ben.stevenson@auckland.ac.nz}

Authors:
\itemize{
  \item Ben Stevenson \email{ben.stevenson@auckland.ac.nz}
  \item Joseph Reps \email{jrep150@aucklanduni.ac.nz}
  \item Lingyu Hao \email{lhao446@aucklanduni.ac.nz}
}

}
\keyword{internal}
