% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{AIC.acre}
\alias{AIC.acre}
\title{Calculate AIC for acre Models}
\usage{
\method{AIC}{acre}(..., k = 2)
}
\arguments{
\item{...}{One or more acre model objects.}

\item{k}{Penalty parameter (default is 2 for AIC).}
}
\value{
A named vector of AIC values or NA if conditions are not met.
}
\description{
Calculate AIC for acre Models
}
