% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{AIC.acre_tmb}
\alias{AIC.acre_tmb}
\title{Calculate AIC for acre_tmb Models}
\usage{
\method{AIC}{acre_tmb}(..., k = 2)
}
\arguments{
\item{...}{One or more acre_tmb model objects.}

\item{k}{Penalty parameter (default is 2 for AIC).}
}
\value{
A named vector of AIC values or NA if conditions are not met.
}
\description{
Calculate AIC for acre_tmb Models
}
\examples{
AIC.acre_tmb(model1, model2, k = 2)
}
