% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.acreboot}
\alias{coef.acreboot}
\title{Title}
\usage{
\method{coef}{acreboot}(
  object,
  types = NULL,
  pars = NULL,
  new.covariates = NULL,
  correct_bias = FALSE,
  ...
)
}
\arguments{
\item{object}{an object generated by the bootstrap function "boot.acre()".}

\item{types}{a character vector, the same as "coef.acre()".}

\item{pars}{a character vector, the same as "coef.acre()".}

\item{new.covariates}{a data frame, the same as "coef.acre()".}

\item{correct_bias}{logical. if TRUE, apply a bias correction method to the bootstrap results; otherwise,
return the original estimation directly. The bias correct method is
hat(coef) - (mean(boot_results) - hat(coef)).}

\item{...}{For S3 compatibility.}
}
\description{
Title
}
