% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{confint.acre}
\alias{confint.acre}
\title{Extract confidence interval for acre.tmb models}
\usage{
\method{confint}{acre}(
  object,
  parm = NULL,
  level = 0.95,
  types = NULL,
  new.covariates = NULL,
  ...
)
}
\arguments{
\item{object}{a fitted model from "fit.acre()".}

\item{parm}{a character vector, the same as "coef.acre()".}

\item{level}{a numeric value indicates the confident level, default is 0.95.}

\item{types}{a character vector, the same as "coef.acre()".}

\item{new.covariates}{a data frame, the same as "coef.acre()".}

\item{...}{For S3 compatibility.}
}
\value{
a matrix
}
\description{
Extract confidence interval for acre.tmb models
}
