% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert.mask}
\alias{convert.mask}
\title{Convert mask object}
\usage{
convert.mask(mask)
}
\arguments{
\item{mask}{an object generated by the function `create.mask()`, if not provided, this function will create it based on
provided `traps` and `control.mask`.}
}
\value{
An object of class \code{mask}.
}
\description{
Converts an \code{acre} mask matrix to a \code{secr} mask
object.
}
\details{
The returned object is suitable for use as the \code{mask}
argument of the function \link{secr.fit}.
}
