% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{create.mask}
\alias{create.mask}
\title{Create mask object}
\usage{
create.mask(traps, buffer, ...)
}
\arguments{
\item{traps}{a matrix or a data frame with two columns or a list of such matrices or data frames for a multi-session model.
Each row in a matrix/data frame provides Cartesian coordinates (in metres) for the location of a detector.
In a list of matrices or data frames, each element of the list corresponds to the detector location of a different session.
If the detector locations stayed the same across several sessions, only one matrix/data frame is required.}

\item{buffer}{a scalar, the furthest distance that a detector could detect (in metres)}

\item{...}{}
}
\description{
Creates a mask object to use with the function read.acre
}
