% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{create_locations_plot_title}
\alias{create_locations_plot_title}
\title{Create a plot subtitle for \link{plot_location_density}}
\usage{
create_locations_plot_title(call_id, animal_id, session)
}
\arguments{
\item{call_id}{an numeric vector indicating which call ID's to calculate
density for.}

\item{animal_id}{an numeric value indicating animal ID. Should only be 
specified if \code{fit} is an \code{animalID} model.}

\item{session}{an integer value indicating the detector array and 
individual(s) to be plotted.}
}
\value{
a string to use as plot subtitle
}
\description{
Create a plot subtitle for \link{plot_location_density}
}
