% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{dist_nearest}
\alias{dist_nearest}
\title{A helper function to obtain the distances of the nearest points from a data frame}
\usage{
dist_nearest(from, to, col_name = "dist_nearest")
}
\arguments{
\item{from}{a matrix or a data frame with columns "x" and "y" contains the coordinates of the start points.}

\item{to}{a matrix or a data frame with columns "x" and "y" contains the coordinates of the end points.}

\item{col_name}{a character, contains the new column name for the distances in the output, default is
dist_nearest.}
}
\value{
a data frame as the same as "from", but with an extra column with assigned column name. For each row,
it contains the distance to the nearest point in the "to" data set.
}
\description{
A helper function to obtain the distances of the nearest points from a data frame
}
