% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations2.R
\name{Dsurface_density}
\alias{Dsurface_density}
\title{Calculate the detection surface density}
\usage{
Dsurface_density(
  fit,
  session = 1,
  mask = get_mask(fit)[[session]],
  new_covariates = NULL
)
}
\arguments{
\item{fit}{an object generated from the model fitting function 
\link{fit.acre} or the bootstrap process \link{boot.acre}.}

\item{session}{an integer value indicating the detector array and 
individual(s) to be plotted.}

\item{mask}{a list or a data frame. Generated by \link{create.mask()}.}

\item{new_covariates}{data.frame; contains any covariates that will be used 
for all extended parameters.}
}
\value{
a data frame with columns \code{x}, \code{y} and \code{Dsurface_dens}
}
\description{
Calculates the detection surface density for a model fit using 
\link{fit.acre}.
}
