% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{esa}
\alias{esa}
\title{Estimated effective sampling areas}
\usage{
esa(fit)
}
\arguments{
\item{fit}{A fitted model from \code{\link[=fit.acre]{fit.acre()}}}
}
\description{
Extracts an estimate and standard error of each session's effective
sampling area.
}
