% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_acre.R
\name{fit.acre}
\alias{fit.acre}
\title{Fitting a spatial capture-recapture model}
\usage{
fit.acre(
  data,
  model = NULL,
  detfn = NULL,
  ss.opts = NULL,
  control.mask = NULL,
  mask = NULL,
  convert.loc2mask = list(),
  tracing = TRUE,
  optim.opts = NULL,
  two.stage = FALSE,
  CL = two.stage
)
}
\arguments{
\item{data}{A data object returned by \code{\link[=read.acre]{read.acre()}}.}

\item{model}{A list with named components. Each component name must
match a parameter name. The component itself must be a \link{formula}
specifying the relationship between covariates and the
parameter. See the section on model specifications below.}

\item{detfn}{A character string specifying the detection function
to be used. Either \code{"hn"} (halfnormal), \code{"hhn"} (hazard
halfnormal), \code{"hr"} (hazard rate), \code{"th"} (threshold), \code{"lth"}
(log-link threshold), or \code{"ss"} (signal strength).  If \code{"ss"} is
used, signal strength information must be included in \code{data}. See
the section below on parameter names for further details.}

\item{ss.opts}{A list with information required to fit models that
include signal strengths as auxiliary detection data. One
component must be named \code{cutoff}, a detection threshold. An
optional component is \code{ss.link}, which specifies the relationship
between distance and the expected received signal strength. See
the section below on signal strength models for further details.}

\item{control.mask}{a list with elements corresponding to the
function \code{create.mask}. If provided, and the argument "mask" is
not provided, it will be used together with "traps" in the "dat"
to expand masks.}

\item{mask}{an object generated by the function \code{create.mask()}, if
not provided, this function will create it based on provided
\code{traps} and \code{control.mask}.}

\item{convert.loc2mask}{a list. When either the arguments of \code{mask}
or \code{control.mask} is provided, the mask will be created by them,
and if there is any parameter extended under mask level, this
argument could be used to control the process of converting
location related covariates to the new mask level data.  For any
details, could refer to the help document in the function
\code{read.acre()}.}

\item{tracing}{A logical value. If \code{TRUE}, information about the
optimiser's journey through the parameter space is printed during
model fitting.}

\item{optim.opts}{A list with optimisation options. See the section
below on optimiser settings.}

\item{two.stage}{A logical value. If \code{TRUE}, computations required
to propagate variance under a two-stage density modelling
approach are carried out.}

\item{CL}{A logical value. If \code{TRUE}, a conditional likelihood
model is fitted, and so density is not estimated.}
}
\description{
Fits an a spatial capture-recapture model to a data object returned by \code{\link[=read.acre]{read.acre()}}.
}
\section{Model specification}{


Specify models using the \code{model} argument.

(More to put in here)
}

\section{Parameter names}{


Parameters fall into three groups:
\itemize{
\item \code{D}, density (in calls or animals per hectare).
\item Detection function parameters.
\item Auxiliary data measurement error parameters.
}

(More to put in here)
}

\section{Optimiser settings}{


The \code{\link[=fit.acre]{fit.acre()}} function fits a spatial capture-recapture model by
numerically maximising the likelihood function with respect to the
model parameters. In almost all cases, the optimisation procedure
will run without a problem, but in rare cases it might help to
adjust how the optimiser works.

The \code{optim.opts} argument allows the user to change some optimiser
settings. The argument must be a list. Two optional components
control parameter-specific optimisation settings:
\itemize{
\item \code{sv}: Overrides default parameter start values.
\item \code{fix}: Holds parameters constant rather than estimating them via
optimisation.
}

These must be named lists, where component names are parameter
names. For \code{sv}, the component is a scalar that provides the
parameter's start value, and for \code{fix} the component is a scalar
that provides the fixed value of the parameter. So, for example,
\code{sv = list(sigma = 200)} will set a start value of 200 m for
\code{sigma}.

It is not possible to control coefficients for covariates. These
always have start values of zero (i.e., no effect of the
covariate), and cannot be fixed. The \code{sv} component can be named
after a parameter that is modelled with covariates. For example, if
we include \code{model = list(D = ~ elevation)}, where \code{elevation} is
height above sea level in metres, then using \code{sv = list(D = 100)}
sets the starting values so that density is spatially homogenous at
100 calls per hectare (or 100 animals per hectare, depending on the
model). The \code{fix} component cannot be used for parameters that are
modeled with covariates.

Three additional components control optimiser behaviour:
\itemize{
\item \code{gr.skip}: A logical value, and \code{FALSE} by default. We use
automatic differentiation to speed up numeric optimisation, but
in some cases this requires large amounts of RAM. Setting
\code{gr.skip} to \code{TRUE} turns off automatic differentiation, which
reduces RAM demands, but, if you have enough RAM, will increase
computation time.
\item \code{scale.covs}: A logical value, and \code{TRUE} by default. If \code{TRUE},
covariates are centred and standardised (i.e., we subtract the
mean, then divide by the standard deviation) prior to
optimisation. This leads to an equivalent model, but optimisation
is typically more stable. Regardless of whether \code{scale.covs} is
\code{TRUE} or \code{FALSE}, we report estimated coefficients and standard
errors on the scale of the original, untransformed covariate.
\item \code{local}: A logical value, and \code{FALSE} by default, that toggles
local integration over activity centres. If \code{TRUE}, the
likelihood will be computed by only integrating over mask points
that are within the buffer distance of all traps that made a
detection. This can speed up optimisation, but makes estimates
more sensitive to the mask buffer setting.
}
}

\section{Signal strength models}{


There are a few options available for signal strength models.

(More to put in here)
}

