% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations2.R
\name{geom_contour_helper}
\alias{geom_contour_helper}
\title{Construct location density plot contours}
\usage{
geom_contour_helper(dens, filled = T, nlevels = 8, prob = F)
}
\arguments{
\item{dens}{a vector of probability densities.}

\item{filled}{a logical value indicating whether the contour should be 
a filled or line contour.}

\item{nlevels}{a numeric value indicating the number of contour lines to
plot.}

\item{prob}{a logical value indicating whether the contours labels should be 
displayed as probabilities or density intervals.}
}
\value{
a list containing ggplot \code{geom_contour} object, as well as a 
   \code{scale_fill_manual} object in the case of filled contours.
}
\description{
A helper function to construct the contours used in 
\link{plot_location_density}
}
