% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{get_dataset_names}
\alias{get_dataset_names}
\title{Get all dataset names.}
\usage{
get_dataset_names()
}
\value{
A character vector containing dataset names.
}
\description{
Retrieves all the dataset names included in the [acre] package
}
\examples{
dataset_names <- get_dataset_names()
}
