% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funs.R
\name{get_par_extend_matrix}
\alias{get_par_extend_matrix}
\title{Calculate and return coefficients for extended models}
\usage{
get_par_extend_matrix(fit, mask, traps, newdata = NULL, warn = T)
}
\arguments{
\item{fit}{}

\item{mask}{}

\item{traps}{}

\item{newdata}{}

\item{select_covariates}{}
}
\value{
a list of (nmask) x (ntrap) columns, for each extended parameter.
         values correspond to estimated parameter value for the given covariate values.
}
\description{
Calculate and return coefficients for extended models over the entire mask.
}
\details{
Returns a list of (nmask) x (ntrap) columns, for each extended parameter.
values correspond to estimated parameter value for either
(a) The covariate values used to fit the model or
(b) The covariate values provided in newdata

Note that because the parameter estimates are calculated for the entire mask,
over every trap, newdata argument will be recycled (or truncated), to the 
correct dimensions.
}
