% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{ihd_ext}
\alias{ihd_ext}
\title{Data to showcase a "ihd_ext" demo}
\format{
a list with all necessary input for the model:
\describe{
\item{detfn}{detechtion function: "hn" - half normal}
\item{model}{a list with the name of coefficients to be modeled with additional covariates
as the elements names, and each element contains a formula for that modeled coefficient.}
\item{traps}{a data frame with the coordinates of the acoustic detectors}
\item{control.mask}{a list with the basic argument "buffer", which is the max detectable distance,
to create the detectable area from the coordinates of the detectors}
\item{trap.cov}{a data frame contains detectors related covariates for the modeled coefficients, it must
contains the indices of detectors, if the detectors are different between any survey sessions,
the indices of sessions should be provided as well}
\item{loc.cov}{a data frame contains location related covariates for the modeled coefficients, it must
contains column 'x' and 'y' as coordinates.}
\item{capt}{a data frame of the capture history, contains columns as follows:\cr
\cr
"session" - the indices of the survey sessions;\cr
"ID" - the indices of the calls been detected;\cr
"occasion" - not been used, could be ignored;\cr
"trap" - the indices of the detectors.\cr
}
\item{fix}{a list, contains the coefficient which been fixed instead of estimated by the model}
}
}
\source{
created from the simulation
}
\usage{
ihd_ext
}
\description{
For the demonstration of the model with inhomogeneous density surface,
and half normal as detection function. One of the coefficients of the detection function, "sigma",
is modeled with additional covariates
}
\keyword{datasets}
