% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{ind_ss}
\alias{ind_ss}
\title{Data to showcase a "ind_ss" demo}
\format{
a list with all necessary input for the model:
\describe{
\item{detfn}{detechtion function: "ss" - signal strength}
\item{model}{a list with the name of coefficients to be modeled with additional covariates
as the elements names, and each element contains a formula for that modeled coefficient.}
\item{traps}{a data frame with the coordinates of the acoustic detectors}
\item{control.mask}{a list with the basic argument "buffer", which is the max detectable distance,
to create the detectable area from the coordinates of the detectors}
\item{trap.cov}{a data frame contains detectors related covariates for the modeled coefficients, it must
contains the indices of detectors, if the detectors are different between any survey sessions,
the indices of sessions should be provided as well}
\item{loc.cov}{a data frame contains location related covariates for the modeled coefficients, it must
contains column 'x' and 'y' as coordinates.}
\item{ss.opts}{a list contains signal strength model related options. Here it contains "cutoff",
which indicates the threshold of a distance of 100\% detection.}
\item{control_create_cpat}{a list with arguments for the function "create.capt()" which converts the data input
to the data to be fed into the model.}
\item{capt}{a data frame of the capture history, contains columns as follows:\cr
\cr
"session" - the indices of the survey sessions;\cr
"ID" - the indices of the calls been detected;\cr
"occasion" - not been used, could be ignored;\cr
"trap" - the indices of the detectors;\cr
"animal_ID" - the indices or identity of the detected individual;\cr
"ss" - the detected signal strength.\cr
}
}
}
\source{
created from the simulation
}
\usage{
ind_ss
}
\description{
For the demonstration of the model with inhomogeneous density surface, signal strength as extra information
and detection function. One of the coefficients of detection function, "b0.ss", is modeled by additional covariates.
The survey is carried with 3 sessions, and each records in the capture history contains identity of the detected individual.
}
\keyword{datasets}
