% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{ind_toa_hhn}
\alias{ind_toa_hhn}
\title{Data to showcase a "ind_toa_hhn" demo}
\format{
a list with all necessary input for the model:
 \describe{
   \item{detfn}{detechtion function: "hhn" - hazard half normal}
   \item{model}{a list with the name of coefficients to be modeled with additional covariates
                           as the elements names, and each element contains a formula for that modeled coefficient.}
   \item{traps}{a data frame with the coordinates of the acoustic detectors}
   \item{control.mask}{a list with the basic argument "buffer", which is the max detectable distance,
                              to create the detectable area from the coordinates of the detectors}
   \item{loc.cov}{a data frame contains location related covariates for the modeled coefficients, it must
                  contains column 'x' and 'y' as coordinates.}
   \item{control_create_cpat}{a list with arguments for the function "create.capt()" which converts the data input
                              to the data to be fed into the model.}
   \item{capt}{a data frame of the capture history, contains columns as follows:\cr
     \cr
     "session" - the indices of the survey sessions;\cr
     "ID" - the indices of the calls been detected;\cr
     "occasion" - not been used, could be ignored;\cr
     "trap" - the indices of the detectors;\cr
     "animal_ID" - the indices or identity of the detected individual;\cr
     "toa" - the time of arrival.\cr
     }
 }
}
\source{
created from the simulation
}
\usage{
ind_toa_hhn
}
\description{
For the demonstration of the model with inhomogeneous density surface, time of arrival as extra information,
and hazard half normal as detection function. The survey is carried with 2 sessions, and each records in the
capture history contains identity of the detected individual.
}
\examples{
o_demo = demo_fit("ind_toa_hhn")
}
\keyword{datasets}
