% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{location_density}
\alias{location_density}
\title{Calculate location density}
\usage{
location_density(
  fit,
  call_id,
  animal_id = NULL,
  session = 1,
  mask = get_mask(fit)[[session]],
  warn_toa = F,
  newdata = NULL
)
}
\arguments{
\item{fit}{an object generated from the model fitting function 
\link{fit.acre} or the bootstrap process \link{boot.acre}.}

\item{call_id}{an numeric vector indicating which call ID's to calculate
density for.}

\item{animal_id}{an numeric value indicating animal ID. Should only be 
specified if \code{fit} is an \code{animalID} model.}

\item{session}{an integer value indicating the detector array and 
individual(s) to be plotted.}

\item{mask}{a list or a data frame. Generated by \link{create.mask()}.}

\item{warn_toa}{a logical value indicates whether to issue warning in the 
case of \code{toa} data with only a single call.}

\item{newdata}{data.frame; contains any covariates that will be used 
for all extended parameters.}
}
\description{
Calculates the estimated density of a location, given its observed capture 
history and auxiliary information.
}
