% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLik.acre}
\alias{logLik.acre}
\title{Extract (negative) log-likelihood for a fitted acre object}
\usage{
\method{logLik}{acre}(object, ...)
}
\arguments{
\item{object}{a fitted model from "fit.acre()".}

\item{...}{For S3 compatibility.}
}
\value{
a numeric value, the negative log-likelihood of the model
}
\description{
Extract (negative) log-likelihood for a fitted acre object
}
