% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLik.acre_tmb}
\alias{logLik.acre_tmb}
\title{Extract (negative) log-likelihood for a fitted acre object}
\usage{
\method{logLik}{acre_tmb}(object, ...)
}
\arguments{
\item{object}{a fitted model from "fit.acre()".}

\item{...}{}
}
\value{
a numeric value, the negative log-likelihood of the model
}
\description{
Extract (negative) log-likelihood for a fitted acre object
}
\examples{
logLik(example.fit)
}
