% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_in_fit.R
\name{make_trace_printer}
\alias{make_trace_printer}
\title{Title}
\usage{
make_trace_printer(
  trace_cols,
  only_improvements = TRUE,
  step = c("euclid", "max", "none")
)
}
\arguments{
\item{trace_cols}{}

\item{only_improvements}{}

\item{step}{}
}
\description{
Title
}
