% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{p.dot}
\alias{p.dot}
\title{Calulating probabilities of detection at mask points}
\usage{
p.dot(fit)
}
\arguments{
\item{fit}{a model object generated by \code{fit.acre}.}
}
\value{
A list with a component for each session, containing a
vector of detection probabilities.
}
\description{
Calculates the probability of a call produced at a mask point being
detected by one or more detectors.
}
