% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyplots.R
\name{plot.acre}
\alias{plot.acre}
\title{Plotting acre model objects}
\usage{
\method{plot}{acre}(x, ...)
}
\arguments{
\item{x}{}

\item{...}{For S3 compatibility.}
}
\description{
Plotting acre model objects
}
