% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyplots.R
\name{plot.ascr_data}
\alias{plot.ascr_data}
\title{Title}
\usage{
\method{plot}{ascr_data}(
  dat,
  types = NULL,
  session = NULL,
  anime = FALSE,
  control = NULL,
  ask = TRUE,
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{dat}{}

\item{types}{a character, "survey", "capt" or "covariates"}

\item{session}{}

\item{anime}{}

\item{control}{a list,}

\item{ask}{}

\item{xlim}{}

\item{ylim}{}

\item{...}{}
}
\description{
Title
}
