% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyplots.R
\name{plot.ascr_tmb}
\alias{plot.ascr_tmb}
\title{Title}
\usage{
\method{plot}{ascr_tmb}(
  fit,
  types = NULL,
  session = NULL,
  anime = FALSE,
  control = NULL,
  ask = TRUE,
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{fit}{}

\item{types}{}

\item{session}{}

\item{anime}{}

\item{control}{}

\item{ask}{}

\item{xlim}{}

\item{ylim}{}

\item{...}{}
}
\description{
Title
}
