% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations2.R
\name{plot_distance_circles}
\alias{plot_distance_circles}
\title{Creates ggplot \code{geom_path} object representing call distance}
\usage{
plot_distance_circles(fit, call_id, animal_id = NULL, session = 1)
}
\arguments{
\item{fit}{an object generated from the model fitting function 
\link{fit.acre} or the bootstrap process \link{boot.acre}.}

\item{call_id}{an numeric vector indicating which call ID's to calculate
density for.}

\item{animal_id}{an numeric value indicating animal ID. Should only be 
specified if \code{fit} is an \code{animalID} model.}

\item{session}{an integer value indicating the detector array and 
individual(s) to be plotted.}
}
\value{
a list containing a ggplot \code{geom_path} object.
}
\description{
Creates circles which have radius equal to the recorded distance for each 
call. The circles are centered around the corresponding triggered traps 
location.
}
