% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations2.R
\name{plot_location_density}
\alias{plot_location_density}
\title{Plot location density}
\usage{
plot_location_density(
  fit,
  call_id = "all",
  animal_id = NULL,
  session = 1,
  mask = get_mask(fit)[[session]],
  new_covariates = NULL,
  plot_types = "combined",
  joint = F,
  nlevels = 8,
  ann = T,
  ask = interactive(),
  display_plot = T
)
}
\arguments{
\item{fit}{an object generated from the model fitting function 
\link{fit.acre} or the bootstrap process \link{boot.acre}.}

\item{call_id}{an numeric vector indicating which call ID's to calculate
density for.}

\item{animal_id}{an numeric value indicating animal ID. Should only be 
specified if \code{fit} is an \code{animalID} model.}

\item{session}{an integer value indicating the detector array and 
individual(s) to be plotted.}

\item{mask}{a list or a data frame. Generated by \link{create.mask()}.}

\item{new_covariates}{data.frame; contains any covariates that will be used 
for all extended parameters.}

\item{plot_types}{a vector which is a subset of 
\code{c("capt", "ss", "dist", "bearing", "toa", "Dsurface", "combined")}}

\item{joint}{a logical value indicating whether to plot the joint density.
Only applicable to animal ID models. Not yet implemented.}

\item{nlevels}{a numeric value indicating the number of contour lines to
plot.}

\item{ann}{a logical value indicating whether to include plot annotation.}

\item{ask}{a logical value indicating whether to prompt user for subsequent
plot.}

\item{display_plot}{a logical value indicating whether to display the plot.}
}
\value{
a ggplot object corresponding to the most recently constructed plot.
}
\description{
A helper function provided to quickly plot location density.
}
\details{
If you want more customization of the plot, please consider using 
\link{location_density} as well as any of plot helper functions:
- \link{plot_traps}
- \link{plot_toa_order}
- \link{plot_bearing_arrows}
- \link{plot_distance_circles}
}
