% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations2.R
\name{plot_ss_indicators}
\alias{plot_ss_indicators}
\title{Creates ggplot \code{geom_point} object representing call signal strength}
\usage{
plot_ss_indicators(fit, call_id, animal_id, session = 1)
}
\arguments{
\item{fit}{an object generated from the model fitting function 
\link{fit.acre} or the bootstrap process \link{boot.acre}.}

\item{call_id}{an numeric vector indicating which call ID's to calculate
density for.}

\item{animal_id}{an numeric value indicating animal ID. Should only be 
specified if \code{fit} is an \code{animalID} model.}

\item{session}{an integer value indicating the detector array and 
individual(s) to be plotted.}
}
\value{
a list containing ggplot \code{geom_point} objects, as well as a 
   \code{guides} object.
}
\description{
Creates blue circles, which get darker based on signal strength intensity.
Not currently implemented as it does not add much to the plotting.
}
