% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations2.R
\name{plot_toa_order}
\alias{plot_toa_order}
\title{Creates ggplot \code{geom_text} annotations of \code{toa} data}
\usage{
plot_toa_order(fit, call_id, animal_id = NULL, session = 1)
}
\arguments{
\item{fit}{an object generated from the model fitting function 
\link{fit.acre} or the bootstrap process \link{boot.acre}.}

\item{call_id}{an numeric vector indicating which call ID's to calculate
density for.}

\item{animal_id}{an numeric value indicating animal ID. Should only be 
specified if \code{fit} is an \code{animalID} model.}

\item{session}{an integer value indicating the detector array and 
individual(s) to be plotted.}
}
\value{
a ggplot \code{geom_text} object.
}
\description{
Annotates traps, in order of the time of arrival of the call. 1 for the first
trap to receive the call, 2 for the second...
}
