% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations2.R
\name{plot_traps}
\alias{plot_traps}
\title{Creates ggplot \code{geom_point} traps of a fitted \code{acre} model}
\usage{
plot_traps(
  fit,
  call_id = NULL,
  animal_id = NULL,
  session = 1,
  circle_traps = F,
  col = "red"
)
}
\arguments{
\item{fit}{an object generated from the model fitting function 
\link{fit.acre} or the bootstrap process \link{boot.acre}.}

\item{call_id}{an numeric vector indicating which call ID's to calculate
density for.}

\item{animal_id}{an numeric value indicating animal ID. Should only be 
specified if \code{fit} is an \code{animalID} model.}

\item{session}{an integer value indicating the detector array and 
individual(s) to be plotted.}

\item{circle_traps}{a logical value indicating whether or not triggered traps 
should be circled. If \code{TRUE}, \code{call_id} (and \code{animal_id} in
the case of animal ID models) must be provided.}
}
\value{
a list containing either 1 or 2 ggplot \code{geom_point} objects, 
   depending on whether \code{circle_traps = T}.
}
\description{
Creates crosses at trap locations. Optionally circles triggered traps for a 
given call.
}
