% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.ascr_boot}
\alias{predict.ascr_boot}
\title{Title}
\usage{
\method{predict}{ascr_boot}(
  fit,
  type = "response",
  newdata = NULL,
  se.fit = TRUE,
  confidence = TRUE,
  level = 0.95,
  realnames = NULL,
  correct_bias = FALSE,
  from_boot = TRUE,
  ...
)
}
\arguments{
\item{fit}{an object generated by the bootstrap function "boot.ascr()".}

\item{type}{a character vector, the same as "predict.ascr_tmb()".}

\item{newdata}{a data frame, the same as "predict.ascr_tmb()".}

\item{se.fit}{a logical value indicates whether to show standard error.}

\item{confidence}{a logical value indicates whether to show confidence interval.}

\item{level}{a numeric value indicates the confident level, default is 0.95.}

\item{realnames}{a character vector, the same as "predict.ascr_tmb()".}

\item{correct_bias}{a logical value indicates whether to apply the bias correction method to the
point estimations and the confidence intervals. The method is described in the
functions "coef.ascr_boot()" and "confint.ascr_boot()".}

\item{from_boot}{a logical value indicates whether to extract standard error and confidence interval
from the bootstrap results.}

\item{...}{}
}
\description{
Title
}
