% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.ascr_tmb}
\alias{predict.ascr_tmb}
\title{Title}
\usage{
\method{predict}{ascr_tmb}(
  fit,
  type = "response",
  newdata = NULL,
  se.fit = TRUE,
  confidence = TRUE,
  level = 0.95,
  realnames = NULL,
  ...
)
}
\arguments{
\item{fit}{a fitted model from "fit.ascr()".}

\item{type}{a character vector, could be either "response" or "link". The "response" correspondence to
the type of "fitted" in the function "coef.ascr_tmb()" and the "link" correspondence to
the type of "linked" in the function "coef.ascr_tmb()".}

\item{newdata}{a data frame, the same as "coef.ascr_tmb()". If there is any extended parameters, and
the corresponding covariates is not provided here, the estimation for such parameters
will not be shown.}

\item{se.fit}{a logical value indicates whether to show standard error.}

\item{confidence}{a logical value indicates whether to show confidence interval.}

\item{level}{a numeric value indicates the confident level, default is 0.95.}

\item{realnames}{a character vector containing any parameter names.}

\item{...}{}
}
\description{
Title
}
