% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.coef_acre_tmb}
\alias{print.coef_acre_tmb}
\title{S3 method for output of coef.acre_tmb and coef.acre_boot}
\usage{
\method{print}{coef_acre_tmb}(x, ...)
}
\arguments{
\item{x}{an object generated by the function "coef.acre_tmb()" or "coef.acre_boot()".}

\item{...}{}
}
\description{
S3 method for output of coef.acre_tmb and coef.acre_boot
}
