% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.std_acre_tmb}
\alias{print.std_acre_tmb}
\title{S3 method for the output of stdEr.acre_tmb}
\usage{
\method{print}{std_acre_tmb}(x, ...)
}
\arguments{
\item{x}{object with class std_acre_tmb}

\item{...}{}
}
\description{
S3 method for the output of stdEr.acre_tmb
}
