% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ascr.R
\name{read.ascr}
\alias{read.ascr}
\title{Title}
\usage{
read.ascr(
  captures,
  traps,
  mask = NULL,
  control_create_mask = list(),
  control_create_capt = list(),
  loc_cov = NULL,
  control_convert_loc2mask = list(),
  session_cov = NULL,
  trap_cov = NULL,
  dist_cov = NULL,
  cue.rates = NULL,
  survey.length = NULL,
  sound.speed = 331,
  unit_convert_dist = 1,
  unit_convert_time = 1,
  ...
)
}
\arguments{
\item{captures}{a data frame, be passed to function "create.capt()", please refer to the description
for that function.}

\item{traps}{a data frame or a list, be passed to function "create.capt()", please refer to the description
for that function.}

\item{mask}{an object generated by the function "create.mask()", if not provided, this function will create it based on 
provided "traps" and "control_create_mask".}

\item{control_create_mask}{a list. It is used together with "traps" to expend masks. Elements should be
corresponding the arguments of the function "create.mask()" excluding "traps".}

\item{control_create_capt}{a list, contains all other arugments of the function "create.capt()" excluding
the "captures" and "traps".}

\item{loc_cov}{a list or a data frame. When a data frame, it must contain columns "x" and "y" and columns
of covariates, so it is a set of locations with covariates values. When a list, each element
is a data frame, so it can contains different sets of locations with covariates values.}

\item{control_convert_loc2mask}{a list with 3 possible elements to control an embedded function inside
this package to convert the location related data provided by "loc_cov"
to masks level using imputation method. The 3 possible elments are:
control_nn2, control_weight and control_char.
=======================================================================
control_nn2: a list contains the arugments for the function RANN::nn2(),
please refer to the help document in that package for further details;
control_weight: a list contains 3 elements control the weighted method.
The function RANN:nn2() gives us a M * k matrix, where M is the number
of mask points, and each row contains k points provided by "loc_cov"
which are nearest to this mask point. For this k nearest points,
we use their values of each covariates to do the imputation to obtain
the estimated value of this mask point. considering the distances
between this mask point and these k nearest points are different,
we need a method to determine their weights. The argument "control_weight"
is used to do so. Its 3 elements are "method", which could be "Shepard"
or "Modified"; "q", which is used to control both methods; "r",
which is used to control method "Modified".
Shepard method: w ~ (1/d)^q
Modified method: modified shepard method, w ~ [max(0, r - d)/(r * d)]^q}

\item{session_cov}{a data frame. It should contain a column of "session" and any session related covariates as
extra columns.}

\item{trap_cov}{a data frame. It is recommended to have columns of "session" and "trap" together, as well as 
any trap related covariates as extra columns. If all sessions share the same set of traps,
the column of "session" could be skipped.}

\item{dist_cov}{a list with named elements. Each element is a data frame contains columns "x" and "y", recording
the edge of an area. The nearest distance from each mask to this area will be calculated and
assigned to the "masks - level" covariates, and the name of this element will be assigned as the
name of this covariate.}

\item{cue.rates}{a numeric vector. contains the recorded cue rates in a series of time periods with identical
length. The length should be equal to the unit length in the "survey.length".}

\item{survey.length}{a numeric vector or a scalar, contains the length of each session. If it is a scalar and
there are multiple sessions, the value will be assigned to all sessions.}

\item{sound.speed}{a scalar, sound speed with unit of "m/s". The default unit of any distance related values
in such as "traps" or "control_create_mask$buffer" or column "dist" in the "captures" are
all in "metre", and default unit of any time related values are "second". If the data are
not in such default units, there are two methods to do the trick. Firstly, input a new value
to "sound.speed" after the manual conversion by the user; secondly, input the arguments of
unit_convert_time and unit_convert_dist, and this function will do the conversion.}

\item{unit_convert_dist}{a numeric value for the conversion between the unit for the data and "metre". If the 
distance unit in data is "km", then assign 1000 here.}

\item{unit_convert_time}{a numeric value for the conversion between the unit for the data and "second". If the 
time unit in data is "millisecond", then assign 0.001 here.}

\item{...}{}
}
\description{
Title
}
