% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_study.R
\name{set_detection_data_NA}
\alias{set_detection_data_NA}
\title{Sets the first n additional detection data information to NA}
\usage{
set_detection_data_NA(capt_data, dataset_name, n_missing)
}
\arguments{
\item{dataset_name}{Name of the data set (must include the additional 
information to be removed in the name)}

\item{n_missing}{Number of data to be set to NA}

\item{data}{Data set to be adjusted}

\item{is_sim}{True if it is simulation data}
}
\description{
Sets the first n additional detection data information to NA
}
