% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.r
\name{show.contour}
\alias{show.contour}
\title{Title}
\usage{
show.contour(
  mask,
  dens,
  nlevels,
  levels,
  prob,
  col = "black",
  lty = 1,
  show.labels,
  plot.contours,
  return.contours = FALSE
)
}
\description{
Title
}
