% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyplots.R
\name{show.detfn}
\alias{show.detfn}
\title{Plot the detection function}
\usage{
show.detfn(
  fit,
  newdata = NULL,
  skip.extend.param = NULL,
  xlim = NULL,
  ylim = NULL,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  col = NULL,
  add = FALSE,
  ...
)
}
\arguments{
\item{fit}{an object generated from the model fitting function \link{fit.acre} or
the bootstrap process \link{boot.acre}.}

\item{newdata}{data.frame; contains any covariates that will be used for all extended parameters (if not be skipped)}

\item{skip.extend.param}{character; skip extended parameter, for skipped extended parameters,
use its intercept as the value for this parameter}

\item{xlim}{a numeric vector with two elements as the range of x-axis.}

\item{ylim}{a numeric vector with two elements as the range of y-axis.}

\item{main}{a string as the main title of the plot.}

\item{xlab}{a string as the sub-title of x-axis.}

\item{ylab}{a string as the sub-title of y-axis.}

\item{col}{a string or a numeric vector indicates the color of the plotted line.}

\item{add}{a logical value indicates whether to add the lines into the existing plot.}

\item{...}{}
}
\description{
Plot the detection function
}
