% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyplots.R
\name{show_detfn}
\alias{show_detfn}
\title{Title}
\usage{
show_detfn(
  fit,
  new_covariates = NULL,
  param_extend_skip = NULL,
  xlim = NULL,
  ylim = NULL,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  col = NULL,
  add = FALSE,
  ...
)
}
\arguments{
\item{fit}{an object generated from the model fitting function "fit.ascr_tmb()" or
the bootstrap process "boot.ascr()".}

\item{new_covariates}{data.frame; contains any covariates will be used for all extended parameters (if not be skipped)}

\item{param_extend_skip}{character; skip extended parameter, for skipped extended parameters,
use its intercept as the value for this parameter}

\item{xlim}{a numeric vector with two elements as the range of x-axis.}

\item{ylim}{a numeric vector with two elements as the range of y-axis.}

\item{main}{a string as the main title of the plot.}

\item{xlab}{a string as the sub-title of x-axis.}

\item{ylab}{a string as the sub-title of y-axis.}

\item{col}{a string or a numeric vector indicates the color of the plotted line.}

\item{add}{a logical value indicates whether to add the lines into the existing plot.}

\item{...}{}
}
\description{
Title
}
