% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyplots.R
\name{show_Dsurf}
\alias{show_Dsurf}
\title{Title}
\usage{
show_Dsurf(
  fit,
  session = NULL,
  show.cv = FALSE,
  new_data = NULL,
  D_cov = NULL,
  xlim = NULL,
  ylim = NULL,
  x_pixels = 50,
  y_pixels = 50,
  zlim = NULL,
  scale = 1,
  plot_contours = TRUE,
  add = FALSE,
  control_convert_loc2mask = NULL,
  arg_col = list(n = 100),
  trap_plot = NULL,
  ...
)
}
\arguments{
\item{fit}{}

\item{session}{}

\item{show.cv}{}

\item{new_data}{}

\item{D_cov}{}

\item{xlim}{}

\item{ylim}{}

\item{x_pixels}{}

\item{y_pixels}{}

\item{zlim}{}

\item{scale}{}

\item{plot_contours}{}

\item{add}{}

\item{arg_col}{}

\item{trap_plot}{}

\item{...}{}
}
\description{
Title
}
