% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcapt.R
\name{sim.data.prepare}
\alias{sim.data.prepare}
\title{animal_ID',}
\usage{
sim.data.prepare(
  detfn,
  param,
  par.extend,
  traps,
  mask,
  survey.length,
  random.location,
  n.sessions
)
}
\description{
animal_ID',
}
