% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_study.R
\name{sim_data}
\alias{sim_data}
\title{Simulate acre-formatted SCR data.}
\usage{
sim_data(
  sim_name,
  n.rand,
  seed = 810,
  suppress_messages = F,
  proportion_missing = 0
)
}
\arguments{
\item{sim_name}{a string denoting type of data to be simulated. Supported 
types can be found using the \code{get_dataset_names()} 
function.}

\item{n.rand}{a numeric value denoting the number of data sets to be 
generated. If n.rand > 1, \code{output$capt} will be a list of 
length \code{n.rand}.}

\item{seed}{a numeric value denoting the random seed, defaults to 810.}

\item{suppress_messages}{a logical value which indicates whether to suppress 
helper messages.}

\item{proportion_missing}{a numeric value used to set a proportion of the 
covariate data generated to NA (currently only 
bearing, distance and toa supported)}
}
\value{
A data frame containing the simulated data, as well as arguments used.
}
\description{
Simulate acre-formatted SCR data.
}
\examples{
simulated_data <- sim_data("bearing_hn", 5)
}
