% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_study.R
\name{sim_study_for_missing_data}
\alias{sim_study_for_missing_data}
\title{Run a simulation study with missing covariate data.}
\usage{
sim_study_for_missing_data(dataset_name, n.rand)
}
\arguments{
\item{dataset_name}{string denoting the name of the dataset to be studied.}

\item{n.rand}{numeric value indicating the number of datasets to be simulated.}
}
\description{
Run a simulation study with missing covariate data.
}
\examples{
sim_study_for_missing_data("bearing_hn", 20)
}
