% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{simple_hr}
\alias{simple_hr}
\title{Data to showcase a "simple_hr" demo}
\format{
a list with all necessary input for the model:
 \describe{
   \item{detfn}{detechtion function: "hr" - hazard rate}
   \item{traps}{a data frame with the coordinates of the acoustic detectors}
   \item{control_create_mask}{a list with the basic argument "buffer", which is the max detectable distance,
                              to create the detectable area from the coordinates of the detectors}
   \item{capt}{a data frame of the capture history, contains columns as follows:\cr
     \cr
     "session" - the indices of the survey sessions;\cr
     "ID" - the indices of the calls been detected;\cr
     "occasion" - not been used, could be ignored;\cr
     "trap" - the indices of the detectors.\cr
     }
   \item{sv}{a list, contains the coefficient which been assigned a start value for modeling}
 }
}
\source{
created from the simulation
}
\usage{
simple_hr
}
\description{
For the demonstration of the model with hazard rate detection function.
}
\examples{
o_demo = demo_fit("simple_hr")
}
\keyword{datasets}
