% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{ss}
\alias{ss}
\title{Data to showcase a "ss" demo}
\format{

}
\source{
created from the simulation
}
\usage{
ss
}
\description{
For the demonstration of the model with signal strength detection function, and signal strength
as additional information.
}
\keyword{datasets}
