% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{ss_density}
\alias{ss_density}
\title{Calculate the signal strength density}
\usage{
ss_density(
  fit,
  call_id,
  animal_id = NULL,
  session = 1,
  mask = get_mask(fit)[[session]],
  newdata = NULL
)
}
\arguments{
\item{fit}{an object generated from the model fitting function 
\link{fit.acre} or the bootstrap process \link{boot.acre}.}

\item{call_id}{an numeric vector indicating which call ID's to calculate
density for.}

\item{animal_id}{an numeric value indicating animal ID. Should only be 
specified if \code{fit} is an \code{animalID} model.}

\item{session}{an integer value indicating the session of detector array and 
individual(s) to be plotted.}

\item{mask}{a list or a data frame. Generated by \link{create.mask()}.}

\item{newdata}{data.frame; contains any covariates that will be used 
for all extended parameters.}
}
\value{
a data frame with columns \code{x}, \code{y} and \code{ss_dens}
}
\description{
Calculates the signal strength density of calls from a model fit using 
\link{fit.acre}.
}
