% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{ss_toa}
\alias{ss_toa}
\title{Data to showcase a "ss_toa" demo}
\format{

}
\source{
created from the simulation
}
\usage{
ss_toa
}
\description{
For the demonstration of the model with signal strength detection function, and signal strength
and time of arriaval as additional information.
}
\examples{
o_demo = demo_fit("ss_toa")
}
\keyword{datasets}
