% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{stdEr.acre_boot}
\alias{stdEr.acre_boot}
\title{Title}
\usage{
\method{stdEr}{acre_boot}(
  object,
  types = NULL,
  pars = NULL,
  new_covariates = NULL,
  from_boot = TRUE,
  show_fixed_par = TRUE,
  ...
)
}
\arguments{
\item{object}{an object generated by the bootstrap function "boot.acre()".}

\item{types}{a character vector, the same as "coef.acre_tmb()".}

\item{pars}{a character vector, the same as "coef.acre_tmb()".}

\item{new_covariates}{a data frame, the same as "coef.acre_tmb()".}

\item{from_boot}{a logical value, the same as "vcov.acre_boot()".}

\item{show_fixed_par}{a logical value, the same as "vcov.acre_tmb()".}

\item{...}{}
}
\description{
Title
}
