% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.ascr_tmb}
\alias{summary.ascr_tmb}
\title{Title}
\usage{
\method{summary}{ascr_tmb}(object, derived_print = FALSE, ...)
}
\arguments{
\item{object}{an object generated from the model fitting function "fit.ascr_tmb()" or
the bootstrap process "boot.ascr()".}

\item{derived_print}{a logical value indicates whether to show the estimations of "esa".}

\item{...}{}
}
\description{
Title
}
