% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{vcov.acre}
\alias{vcov.acre}
\title{Extract variance covariance matrix of the estimated parameters from acre models}
\usage{
\method{vcov}{acre}(
  object,
  types = NULL,
  pars = NULL,
  new.covariates = NULL,
  show_fixed_par = TRUE,
  ...
)
}
\arguments{
\item{object}{a fitted model from fit.acre().}

\item{types}{a character vector, the same as "coef.acre()".}

\item{pars}{a character vector, the same as "coef.acre()".}

\item{new.covariates}{a data frame, the same as "coef.acre()".}

\item{show_fixed_par}{a logical value. To control whether to include the fixed parameters in the covariance matrix.
It is TRUE by default.}

\item{...}{For S3 compatibility.}
}
\value{
a list with matrices as its elements if multiple 'types'. a matrix if only one 'types'.
}
\description{
Extract variance covariance matrix of the estimated parameters from acre models
}
