% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_in_fit.R
\name{with_trace_printer}
\alias{with_trace_printer}
\title{Title}
\usage{
with_trace_printer(fn, printer, par_names)
}
\arguments{
\item{fn}{}

\item{printer}{}

\item{par_names}{}
}
\description{
Title
}
