% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot.acre}
\alias{boot.acre}
\title{Bootstrapping a fitted \code{acre} model}
\usage{
boot.acre(fit, N = 30, n.cores = 1, infotypes = NULL, seed = NULL)
}
\arguments{
\item{fit}{a model object generated by \code{fit.acre}.}

\item{N}{a numerical value indicates the number of iterations of the bootstrap process.}

\item{n.cores}{a numerical value indicates the number of cores to be used. Currently this function doesn't support
parallel computing, so it is fixed to be 1.}

\item{infotypes}{a character vector containing the infotypes will be simulated in the additional bootstraps;
or a list with each element as a character vector, used for doing multiple additional
bootstraps with different settings of infotypes.}

\item{seed}{a numerical value indicates the seed for the randomness.}
}
\description{
Bootstrapping a fitted \code{acre} model
}
