% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{create.mask}
\alias{create.mask}
\title{Create mask object}
\usage{
create.mask(traps, buffer, ...)
}
\arguments{
\item{traps}{a data frame or a list. contains the coordinates of the detectors}

\item{buffer}{a scaler, the furthest distance that a detector could detect}

\item{...}{}
}
\description{
Creates a mask object to use with the function
}
