% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{Dsurface_density}
\alias{Dsurface_density}
\title{Calculate the detection surface density}
\usage{
Dsurface_density(fit, session = 1, mask = get_mask(fit)[[session]])
}
\arguments{
\item{fit}{an object generated from the model fitting function 
\link{fit.acre} or the bootstrap process \link{boot.acre}.}

\item{session}{an integer value indicating the session of detector array and 
individual(s) to be plotted.}

\item{mask}{a list or a data frame. Generated by \link{create.mask()}.}
}
\value{
a data frame with columns \code{x}, \code{y} and \code{Dsurface_density}
}
\description{
Calculates the detection surface density for a model fit using 
\link{fit.acre}.
}
\details{
See: https://research-repository.st-andrews.ac.uk/bitstream/handle/10023/18233/BenStevensonPhDThesis.pdf
(Equation 1.4)
}
