% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_acre.R
\name{fit.acre}
\alias{fit.acre}
\title{Fitting acoustic SCR models}
\usage{
fit.acre(
  dat,
  model = NULL,
  detfn = NULL,
  sv = NULL,
  bounds = NULL,
  fix = NULL,
  ss.opts = NULL,
  control.mask = NULL,
  mask = NULL,
  convert.loc2mask = list(),
  is_scale = TRUE,
  model.link = NULL,
  local = FALSE,
  tracing = TRUE,
  gr_skip = FALSE,
  sv_link = NULL
)
}
\arguments{
\item{dat}{an object generated by the function `read.acre()`}

\item{model}{a list with named elements. Each element contains the formula for the parameter indicated by the name of
this element. For example, model = list(D = ~x1+x2, sigma = ~x3+x4)}

\item{detfn}{character string specifying the detection function to be used. One of "hn` (halfnormal),
`hhn` (hazard halfnormal), `hr` (hazard rate), `th` (threshold), `lth` (log-link threshold), or `ss` (signal strength).
If `ss` is used, signal strength information must be provided in capt.}

\item{sv}{a named list of parameter start values. Element names are parameter names, and each element is a start value for the associated parameter.
See 'Details' for further information on the parameters to be fitted.}

\item{bounds}{A named list of parameter bounds. Element names are parameter names, and each element is a vector of length two, specifying the upper and
lower bounds for the associated parameter.}

\item{fix}{a named list of fixed parameter values. Element names are parameter names to be fixed, and each element is the fixed value for the associated parameter.
If a parameter is fixed, it will not be estimated by the model.}

\item{ss.opts}{a list. It could contain 3 elements related to signal strength model.
cutoff - the threshold of signal strength could be detected,
         must be included if signal strength is provided.
ss.link - the link function of signal strength, could be `identity`, `log` or `spherical`,
          by default, it is `identity`.
directional - a logical value, indicate whether to apply signal strength directional model.}

\item{control.mask}{a list with elements corresponding to the function `create.mask`. If provided, and the argument "mask" is not provided, it will be used together
with "traps" in the "dat" to expand masks.}

\item{mask}{an object generated by the function `create.mask()`, if not provided, this function will create it based on
provided `traps` and `control.mask`.}

\item{convert.loc2mask}{a list. When either the arguments of `mask` or `control.mask` is provided, the mask
will be created by them, and if there is any parameter extended under mask level, this argument
could be used to control the process of converting location related covariates to the new mask level data.
For any details, could refer to the help document in the function `read.acre()`}

\item{is_scale}{a logical value. Indicate whether to standardize the numerical covariates for the extended parameters, it is `TRUE` by default.}

\item{model.link}{a list with named elements. Each element contains the link for the parameter indicated by the name of this element.}

\item{local}{a logical value. FALSE by default. If TRUE, the model will only integrate the masks within the buffer distance for all traps.}

\item{tracing}{a logical value. TRUE by default, an indicator of showing the tracing information. (seems not work as expected.)}

\item{gr_skip}{a logical value. FALSE by default. If TRUE, TMB model will skip the process of generating automatic derivative functions,
which will use less RAM, but the optimization process will consume more time.}

\item{sv_link}{a list; this is mostly for development purpose, not recommended to use}
}
\description{
Fits an acoustic SCR model. Parameter estimation is done by maximum likelihood through an AD Model Builder (ADMB) executable.
}
