% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_acre.R
\name{fit.acre}
\alias{fit.acre}
\title{Title}
\usage{
fit.acre(
  dat,
  par_extend_model = NULL,
  detfn = NULL,
  sv = NULL,
  bounds = NULL,
  fix = NULL,
  ss.opts = NULL,
  control_create_mask = NULL,
  mask = NULL,
  control_convert_loc2mask = list(),
  is_scale = TRUE,
  par_extend_link = NULL,
  local = FALSE,
  tracing = TRUE,
  gr_skip = FALSE,
  sv_link = NULL
)
}
\arguments{
\item{dat}{an object generated by the function "read.acre()"}

\item{par_extend_model}{a list with named elements. Each element contains the formula for the parameter indicated by the name of
this element. For example, par_extend_model = list(D = ~x1+x2, sigma = ~x3+x4)}

\item{detfn}{a string indicating the detection function. Currently, 'hn', 'hhn', 'hr', 'th', 'lth', 'ss'
are supported.}

\item{sv}{a list. Each element should be a named element with the name of the parameter and the
start value of it.}

\item{bounds}{a list. Each element should be a named element with the name of the parameter and the
range of it. The range is expressed by a vector with two numbers: (lower limit, upper limit).}

\item{fix}{a list. Each element should be a named element with the name of the parameter and the
fixed value of it. If a parameter is fixed, it will not be estimated by the model.}

\item{ss.opts}{a list. It could contain 3 elements related to signal strength model.
cutoff - the threshold of signal strength could be detected,
         must be included if signal strength is provided.
ss.link - the link function of signal strength, could be "identity", "log" or "spherical",
          by default, it is "identity".
directional - a logical value, indicate whether to apply signal strength directional model.}

\item{control_create_mask}{a list. If it is provided, and the argument "mask" is not provided, it will be used together
with "traps" in the "dat" to expend masks. Elements should be corresponding the arguments of
the function "create.mask()".}

\item{mask}{an object generated by the function "create.mask()", if not provided, this function will create it based on
provided "traps" and "control_create_mask".}

\item{control_convert_loc2mask}{a list. When either the arguments of "mask" or "control_create_mask" is provided, the mask
will be created by them, and if there is any parameter extended under mask level, this argument
could be used to control the process of converting location related covariates to the new mask level data.
For any details, could refer to the help document in the function "read.acre()"}

\item{is_scale}{a logical value. Indicate whether to standardize the numerical covariates for the extended parameters, it is "TRUE" by default.}

\item{par_extend_link}{a list with named elements. Each element contains the link for the parameter indicated by the name of this element.}

\item{local}{a logical value. FALSE by default. If TRUE, the model will only integrate the masks within the buffer distance for all traps.}

\item{tracing}{a logical value. TRUE by default, an indicator of showing the tracing information. (seems not work as expected.)}

\item{gr_skip}{a logical value. FALSE by default. If TRUE, TMB model will skip the process of generating automatic derivative functions,
which will use less RAM, but the optimization process will consume more time.}

\item{sv_link}{a list; this is mostly for development purpose, not recommended to use}
}
\description{
Title
}
