% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_acre.R
\name{fit_og}
\alias{fit_og}
\title{Original fit function from ascr}
\usage{
fit_og(
  capt,
  traps,
  mask,
  detfn = NULL,
  sv = NULL,
  bounds = NULL,
  fix = NULL,
  ss.opts = NULL,
  cue.rates = NULL,
  survey.length = NULL,
  sound.speed = 331,
  local = FALSE,
  par.extend = NULL,
  tracing = TRUE,
  gr.skip = FALSE,
  sv.link = NULL,
  CL = FALSE,
  ...
)
}
\arguments{
\item{capt}{a list or a data frame with capture data. Columns named \code{session}, \code{ID} and \code{trap} are required, with
each row being regarded as one detection. Extra information can be provided as columns \code{bearing}, \code{dist}
,\code{ss}, \code{toa}.  \code{animal_ID} could be provided if individuals could be distinguished from their acoustic detection.
In addition, any individuals from different sessions will be treated as independent ones.
The default unit for any distance related value should be 'meter', and for time related value should be 'seconds'..}

\item{traps}{a data frame or a list, the same as the argument 'trap' in function \code{create.capt()}, please refer to the description
for that function.}

\item{mask}{a list or a data frame. Generated by the function \code{create.mask()}.}

\item{detfn}{same as the argument with the same name in the function \code{read.acre()}.}

\item{sv}{same as the argument with the same name in the function \code{read.acre()}.}

\item{bounds}{same as the argument with the same name in the function \code{read.acre()}.}

\item{fix}{same as the argument with the same name in the function \code{read.acre()}.}

\item{ss.opts}{same as the argument with the same name in the function \code{read.acre()}.}

\item{cue.rates}{same as the argument with the same name in the function \code{read.acre()}.}

\item{survey.length}{same as the argument with the same name in the function \code{read.acre()}.}

\item{sound.speed}{same as the argument with the same name in the function \code{read.acre()}.}

\item{local}{same as the argument with the same name in the function \code{fit.acre()}.}

\item{par.extend}{a list. contains elements of \code{model}, \code{data}, \code{scale} and \code{link}.
model: the same as argument \code{model}  in the function \code{fit.acre()};
data: a list, could contain elements of \code{session}, \code{trap}, \code{mask}. Each of
them should be a data frame contains the corresponding information
of sessions, traps or masks and the corresponding covariates;
scale: the same as argument \code{is.scale}  in the function \code{fit.acre()};
link: the same as argument \code{model.link}  in the function \code{fit.acre()}.}

\item{tracing}{same as the argument with the same name in the function \code{fit.acre()}.}

\item{gr.skip}{same as the argument with the same name in the function \code{fit.acre()}.}

\item{sv.link}{a list; this is mostly for development purpose, not recommended to use.}

\item{CL}{a logical value. FALSE by default. If TRUE, fit a conditional likelihood model, ignoring
density parameter estimation.}
}
\description{
Original fit function from ascr
}
\keyword{'ID',}
\keyword{'animal_ID'}
\keyword{and}
\keyword{animal_ID'}
\keyword{are}
\keyword{be}
\keyword{but}
\keyword{combination}
\keyword{contains}
\keyword{data.ID_mask'}
\keyword{design}
\keyword{different}
\keyword{in}
\keyword{internal}
\keyword{make}
\keyword{matrices}
\keyword{may}
\keyword{modify}
\keyword{of}
\keyword{or}
\keyword{order}
\keyword{order,}
\keyword{right}
\keyword{same}
\keyword{so}
\keyword{sure}
\keyword{the}
\keyword{them}
\keyword{they}
\keyword{to}
\keyword{we}
