% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ascr.R
\name{fit_og}
\alias{fit_og}
\title{Title}
\usage{
fit_og(
  capt,
  traps,
  mask,
  detfn = NULL,
  sv = NULL,
  bounds = NULL,
  fix = NULL,
  ss.opts = NULL,
  cue.rates = NULL,
  survey.length = NULL,
  sound.speed = 331,
  local = FALSE,
  par.extend = NULL,
  tracing = TRUE,
  gr_skip = FALSE,
  sv_link = NULL,
  ...
)
}
\arguments{
\item{capt}{an object generated by the function "create.capt()".}

\item{traps}{a data frame or a list, the same as the argument 'trap' in function "create.capt()", please refer to the description
for that function.}

\item{mask}{a list or a data frame. Generated by the function "create.mask()".}

\item{detfn}{same as the argument with the same name in the function "read.ascr()".}

\item{sv}{same as the argument with the same name in the function "read.ascr()".}

\item{bounds}{same as the argument with the same name in the function "read.ascr()".}

\item{fix}{same as the argument with the same name in the function "read.ascr()".}

\item{ss.opts}{same as the argument with the same name in the function "read.ascr()".}

\item{cue.rates}{same as the argument with the same name in the function "read.ascr()".}

\item{survey.length}{same as the argument with the same name in the function "read.ascr()".}

\item{sound.speed}{same as the argument with the same name in the function "read.ascr()".}

\item{local}{same as the argument with the same name in the function "fit.ascr()".}

\item{par.extend}{a list. contains elements of "model", "data", "scale" and "link".
model: the same as argument "par_extend_model"  in the function "fit.ascr()";
data: a list, could contain elements of "session", "trap", "mask". Each of
      them should be a data frame contains the corresponding information
      of sessions, traps or masks and the corresponding covariates;
scale: the same as argument "is_scale"  in the function "fit.ascr()";
link: the same as argument "par_extend_link"  in the function "fit.ascr()".}

\item{tracing}{same as the argument with the same name in the function "fit.ascr()".}

\item{gr_skip}{same as the argument with the same name in the function "fit.ascr()".}

\item{sv_link}{a list; this is mostly for development purpose, not recommended to use.}

\item{...}{}
}
\description{
Title
}
\examples{
animal_ID' or 'ID', so we modify them to make sure design matrices are in right order
data.ID_mask' contains the same combination of 'animal_ID' and 'ID', but they may be in different order,
}
