% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations2.R
\name{get_density_breaks_from_probability}
\alias{get_density_breaks_from_probability}
\title{Calculate density breaks for given probabilities}
\usage{
get_density_breaks_from_probability(dens, probability_breaks)
}
\arguments{
\item{dens}{a vector of probability densities.}

\item{probability_breaks}{the probabilities at which to calculate density 
breaks.}
}
\description{
Calculates breaks for a given density map from provided probability bins.
}
