% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funs.R
\name{get_par_extend_matrix}
\alias{get_par_extend_matrix}
\title{Calculate and return coefficients for extended models}
\usage{
get_par_extend_matrix(fit, mask, traps, newdata = NULL, warn = T)
}
\arguments{
\item{fit}{an object generated from the model fitting function 
\link{fit.acre} or the bootstrap process \link{boot.acre}.}

\item{mask}{a list or a data frame. Generated by \link{create.mask()}.}

\item{traps}{data frame with numeric columns x and y, corresponding to trap
coordinates.}

\item{newdata}{data.frame; contains any covariates that will be used 
for all extended parameters.}

\item{warn}{Logical; warn users if they have not passed all covariates in 
model specification.}
}
\value{
a list of (nmask) x (ntrap) columns, for each extended parameter.
         values correspond to estimated parameter value for the given covariate values.
}
\description{
Returns a list of (nmask) x (ntrap) matrices, for each extended parameter.
values correspond to estimated parameter value for either
(a) The covariate values used to fit the model or
(b) The covariate values provided in newdata
}
\details{
Each row corresponds to a trap, and each column a mask point.
EXAMPLE: let a model have have `sigma` mask-level extended by brand 
(2 traps, 1 "sony" and 1 "panasonic"), and also be mask level extended, say 
with covariate "noise". Then element [1, 1] of the returned `sigma` element,
corresponds to the coefficient of `sigma` with "sony" brand, and noise at
mask point 1. [2, 1] to that of "panasonic" brand, and noise at mask point 1.

Note that because the parameter estimates are calculated for the entire mask,
over every trap, newdata argument will be recycled (or truncated), to the 
correct dimensions.
}
\examples{
# Calculates the parameter values for each mask_point - trap combination
ext_par_matrices <- get_par_extend_matrix(fit, mask, traps)
}
