% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{ind_ss}
\alias{ind_ss}
\title{Data to showcase a "ind_ss" demo}
\format{

}
\source{
created from the simulation
}
\usage{
ind_ss
}
\description{
For the demonstration of the model with inhomogeneous density surface, signal strength as extra information
and detection function. One of the coefficients of detection function, "b0.ss", is modeled by additional covariates.
The survey is carried with 3 sessions, and each records in the capture history contains identity of the detected individual.
}
\examples{
o_demo = demo_fit("ind_ss")
}
\keyword{datasets}
