% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyplots.R
\name{plot.acre_data}
\alias{plot.acre_data}
\title{Plotting acre data}
\usage{
\method{plot}{acre_data}(x, ...)
}
\arguments{
\item{x}{}

\item{...}{For S3 compatibility.}
}
\description{
Plotting acre data
}
