% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_acre.R
\name{read.acre}
\alias{read.acre}
\title{Formatting capture and trap data for model fitting}
\usage{
read.acre(
  captures,
  traps,
  mask = NULL,
  control.mask = list(),
  control.capt = list(),
  loc.cov = NULL,
  time.loc.cov = NULL,
  convert.loc2mask = list(),
  session.cov = NULL,
  trap.cov = NULL,
  dist.cov = NULL,
  cue.rates = NULL,
  survey.length = NULL,
  sound.speed = 331,
  convert.dist.unit = 1,
  convert.time.unit = 1,
  ...
)
}
\arguments{
\item{captures}{a data frame with capture data. Columns named `session`, `ID` and `trap` are required, with
each row being regarded as one detection. Extra information can be provided as columns `bearing`, `dist`
,`ss`, `toa`, `animal_ID`. See 'Captures argument'.}

\item{traps}{a matrix or a data frame with two columns or a list of such matrices or data frames for a multi-session model.
Each row in a matrix/data frame provides Cartesian coordinates (in metres) for the location of a detector.
In a list of matrices or data frames, each element of the list corresponds to the detector location of a different session.
If the detector locations stayed the same across all sessions, only one matrix/data frame is required.}

\item{mask}{an object generated by the function `create.mask()`, if not provided, this function will create the mask based on
provided `traps` and `control.mask`.}

\item{control.mask}{a list. It is used together with `traps` to expend masks. Elements should be
corresponding the arguments of the function `create.mask()` excluding `traps`. e.g. list(buffer = 100)}

\item{control.capt}{a list, contains all other arguments of the function `create.capt()` excluding
the `captures` and `traps`. e.g. list(n.sessions = 5, )}

\item{loc.cov}{a list or a data frame. When a data frame, it must contain columns `x` and `y` and columns
of covariates, so it is a set of locations with covariates values. When a list, each element
is a data frame, so it can contains different sets of locations with covariates values.}

\item{time.loc.cov}{a list or a data frame. It must contain columns `x`, `y` and `time`, and columns of covariates
related to location but could change over time. When this argument is provided, please also
fill in the `time` as a covariate in the argument `session.cov` to help with identifying
the time for each session to allocate these covariates.}

\item{convert.loc2mask}{a list with 3 possible elements to control an embedded function inside
this package to convert the location related data provided by `loc.cov`
to masks level using imputation method. The 3 possible elements are:
control_nn2, control_weight and control_char.
=======================================================================
control_nn2: a list contains the arguments for the function RANN::nn2(),
please refer to the help document in that package for further details;
control_weight: a list contains 3 elements control the weighted method.
The function RANN:nn2() gives us a M * k matrix, where M is the number
of mask points, and each row contains k points provided by `loc.cov`
which are nearest to this mask point. For this k nearest points,
we use their values of each covariates to do the imputation to obtain
the estimated value of this mask point. considering the distances
between this mask point and these k nearest points are different,
we need a method to determine their weights. The argument `control_weight`
is used to do so. Its 3 elements are `method`, which could be `Shepard`
or `Modified`; `q`, which is used to control both methods; `r`,
which is used to control method `Modified`.
Shepard method: w ~ (1/d)^q
Modified method: modified shepard method, w ~ [max(0, r - d)/(r * d)]^q}

\item{session.cov}{a data frame. It should contain a column named `session` and any session related covariates as
extra columns.}

\item{trap.cov}{a data frame. It is recommended to have columns of `session` and `trap` together, as well as
any trap related covariates as extra columns. If all sessions share the same set of traps,
the column of `session` could be skipped.}

\item{dist.cov}{a list with named elements. Each element is a data frame contains columns `x` and `y`, recording
the edge of an area. The nearest distance from each mask to this area will be calculated and
assigned to the `masks - level` covariates, and the name of this element will be assigned as the
name of this covariate.}

\item{cue.rates}{a numeric vector. contains the recorded cue rates in a series of time periods with identical
length. The length should be equal to the unit length in the `survey.length`.}

\item{survey.length}{a numeric vector or a scalar, contains the length of each session. If it is a scalar and
there are multiple sessions, the value will be assigned to all sessions.}

\item{sound.speed}{a scalar, sound speed with unit of `m/s`. The default unit of any distance related values
in such as `traps` or `control.mask$buffer` or column `dist` in the `captures` are
all in `metre`, and default unit of any time related values are `second`. If the data are
not in such default units, there are two methods to do the trick. Firstly, input a new value
to `sound.speed` after the manual conversion by the user; secondly, input the arguments of
convert.time.unit and convert.dist.unit, and this function will do the conversion.}

\item{convert.dist.unit}{a numeric value for the conversion between the unit for the data and `metre`. If the
distance unit in data is `km`, then assign 1000 here.}

\item{convert.time.unit}{a numeric value for the conversion between the unit for the data and `second`. If the
time unit in data is `millisecond`, then assign 0.001 here.}

\item{...}{}
}
\description{
Reads and formats capture and trap data so it can be passed to fit.acre.
}
