% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_acre.R
\name{read.acre}
\alias{read.acre}
\title{Combining all data for model fitting}
\usage{
read.acre(
  captures,
  traps,
  mask = NULL,
  control.mask = list(),
  control.capt = list(),
  loc.cov = NULL,
  time.loc.cov = NULL,
  convert.loc2mask = list(),
  session.cov = NULL,
  trap.cov = NULL,
  dist.cov = NULL,
  cue.rates = NULL,
  survey.length = NULL,
  sound.speed = 331,
  convert.dist.unit = 1,
  convert.time.unit = 1,
  ...
)
}
\arguments{
\item{captures}{A data frame with detection data. Further details are
available below.}

\item{traps}{A matrix or a data frame with two columns, or a list of such
matrices or data frames for a multi-session model. Each row provides the x-
and y-coordinates of a detector location, in metres. For multi-session
models, each component of the list provides detector locations for one
session.}

\item{mask}{Optional. A mask object. If \code{NULL}, a mask object will be created
automatically based on the \code{traps} and \code{control.mask} arguments.}

\item{control.mask}{A list specifying additional arguments for
\code{\link[=create.capt]{create.capt()}}, other than \code{traps} that are used to create a mask object.}

\item{control.capt}{a list, contains all other arguments of the function
\code{create.capt()} excluding the \code{captures} and \code{traps}. e.g. list(n.sessions =
5, )}

\item{loc.cov}{a list or a data frame. When a data frame, it must contain
columns \code{x} and \code{y} and columns of covariates, so it is a set of locations
with covariates values. When a list, each element is a data frame, so it can
contains different sets of locations with covariates values.}

\item{time.loc.cov}{a list or a data frame. It must contain columns \code{x}, \code{y}
and \code{time}, and columns of covariates related to location but could change
over time. When this argument is provided, please also fill in the \code{time} as
a covariate in the argument \code{session.cov} to help with identifying the time
for each session to allocate these covariates.}

\item{convert.loc2mask}{a list with 3 possible elements to control an embeddedfunction inside this package to convert the location related data providedby loc.cov to masks level using imputation method. The 3 possible elementsare: control_nn2, control_weight and control_char.

control_nn2: a list contains the arguments for the function RANN::nn2(),
please refer to the help document in that package for further details;
control_weight: a list contains 3 elements control the weighted method. The
function RANN:nn2() gives us a M * k matrix, where M is the number of mask
points, and each row contains k points provided by \code{loc.cov} which are
nearest to this mask point. For this k nearest points, we use their values
of each covariates to do the imputation to obtain the estimated value of
this mask point. considering the distances between this mask point and these
k nearest points are different, we need a method to determine their weights.
The argument \code{control_weight} is used to do so. Its 3 elements are \code{method},
which could be \code{Shepard} or \code{Modified}; \code{q}, which is used to control both
methods; \code{r}, which is used to control method \code{Modified}. Shepard method: w
~ (1/d)^q Modified method: modified shepard method, w ~ \link[=max(0, r - d)/(r *
 d)]{max(0, r - d)/(r * d)}^q}

\item{session.cov}{a data frame. It should contain a column named \code{session}
and any session related covariates as extra columns.}

\item{trap.cov}{a data frame. It is recommended to have columns of \code{session}
and \code{trap} together, as well as any trap related covariates as extra
columns. If all sessions share the same set of traps, the column of
\code{session} could be skipped.}

\item{dist.cov}{a list with named elements. Each element is a data frame
containing columns \code{x} and \code{y}, recording the edge of an area. The shortest
distance from each mask to this area will be calculated and assigned to the
\code{masks - level} covariates, and the name of this element will be assigned as
the name of this covariate.}

\item{cue.rates}{a numeric vector. Contains the recorded cue rates in a series
of time periods with identical length. A vector of call rates collected
independently ofthe main acoustic survey. This must be measured in calls per
unit time, where the time units are equivalent to those used by
\code{survey.length}. For example, if the survey was 30 minutes long, the
cue rates must be provided in cues per minute if \code{survey.length = 30},
but in cues per hour if \code{survey.length = 0.5}.}

\item{survey.length}{a numeric vector or a scalar, contains the length of each
session. If it is a scalar and there are multiple sessions, the value will
be assigned to all sessions.}

\item{sound.speed}{a scalar, the speed of sound in \code{metres} per \code{second}.
Defaults to 330, approximate speed of sound in air.}

\item{convert.dist.unit}{a numeric value for the conversion between the unit
for the data and \code{metre}. If the distance unit in data is \code{km}, then assign
1000 here.}

\item{convert.time.unit}{a numeric value for the conversion between the unit
for the data and \code{second}. If the time unit for the data is \code{millisecond},
then assign 0.001 here.}

\item{...}{}
}
\description{
Combines detection data, detector locations, and covariate data into a single
object. The resulting object can be used for model fitting with
\code{\link[=fit.acre]{fit.acre()}}.
}