% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcapt.R
\name{sim.capt}
\alias{sim.capt}
\title{Simulating SCR data}
\usage{
sim.capt(
  fit,
  detfn,
  param,
  model = NULL,
  traps,
  control.mask = list(),
  session.cov = NULL,
  trap.cov = NULL,
  loc.cov = NULL,
  dist.cov = NULL,
  time.loc.cov = NULL,
  convert.loc2mask = list(),
  survey.length = NULL,
  ss.opts = NULL,
  cue.rates = NULL,
  n.sessions = NULL,
  n.rand = 1,
  random.location = FALSE,
  sound.speed = 331
)
}
\arguments{
\item{fit}{an object generated from the model fitting function "fit.acre()" or
the bootstrap process "boot.acre()". If \code{fit} is provided, then all other
parameters will be taken from the fitted object, so no other parameters need to be provided.}

\item{detfn}{character string specifying the detection function to be used. One of "hn\verb{(halfnormal),}hhn\verb{(hazard halfnormal),}hr\verb{(hazard rate),}th\verb{(threshold),}lth\verb{(log-link threshold), or}ss\verb{(signal strength). If}ss` is used, signal strength information must be provided in capt.}

\item{param}{A named list. Component names are parameter names, and
each component is the value of the associated parameter. A
value for the parameter \code{D}, animal density (or call
density, if it an acoustic survey) must always be provided,
along with values for parameters associated with the chosen
detection function and additional information type(s).}

\item{model}{a list with named elements. Each element contains the formula for the parameter indicated by the name of
this element. For example, model = list(D = ~x1+x2, sigma = ~x3+x4)}

\item{traps}{A matrix or a data frame with two columns, or a list
of such matrices or data frames for a multi-session model. Each
row provides the x- and y-coordinates of a detector location, in
metres. For multi-session models, each component of the list
provides detector locations for one session.}

\item{control.mask}{A list specifying arguments for
\code{\link[=create.mask]{create.mask()}}, other than \code{traps}, that are used to create a
mask object. The \code{buffer} should be provided, and, optionally,
the \code{spacing}.}

\item{session.cov}{A data frame containing session covariates. It
must contain a column \code{session}, and additional columns for the
session-level covariates.  If spatiotemporal covariates are
included using \code{time.loc.cov}, then a column \code{time} must be
included, indicating when the session took place.}

\item{trap.cov}{A data frame containing trap covariates. It must
contain a column \code{trap} and additional columns for the trap-level
covariates. The column \code{session} must be included for
multisession data.}

\item{loc.cov}{A data frame, or a list of data frames, containing
spatial covariates. Data frames must contain columns \code{x} and \code{y}
for x- and y-coordinates, in metres, and additional columns for
spatial covariates measured at these locations. Missing \code{NA}
values are allowed. Spatial interpolation is performed to produce
covariate values at the mask point locations; see the section on
spatial covariates below.}

\item{dist.cov}{A list containing locations of features from which
distances are calculated, and can be used as spatial
covariates. Each component must be named after a feature, with a
data frame containing columns \code{x} and \code{y}, recording the the
location of a feature.}

\item{time.loc.cov}{A list, or list of data frames, containing
spatial covariates that change over time. The data frames must
contain columns \code{x}, \code{y} and \code{time}, and additional columns for
the spatiotemporal covariates.  When these spatiotemporal
covariates are provided, the column \code{time} must also appear in
\code{session.cov}.}

\item{convert.loc2mask}{A list to control the spatial interpolation
method used to compute covariate values for mask locations based
on data provided in \code{loc.cov} and \code{time.loc.cov}. See the section
on spatial covariates below.}

\item{survey.length}{A numeric vector or a scalar, containing the
length of each session. If it is a scalar and there are multiple
sessions, the value will be assigned to all sessions.}

\item{ss.opts}{a list. It could contain 3 elements related to signal strength model.
cutoff - the threshold of signal strength could be detected,
must be included if signal strength is provided.
ss.link - the link function of signal strength, could be \code{identity}, \code{log} or \code{spherical},
by default, it is \code{identity}.
directional - a logical value, indicate whether to apply signal strength directional model.}

\item{cue.rates}{A numeric vector containing the recorded cue rates
of individuals, collected separately to the SCR survey. See the
section on cue rate data, below.}

\item{n.sessions}{An integer for the number of sessions.}

\item{sound.speed}{A scalar, the speed of sound in metres per
second. This argument is only used when time-of-arrival data are
included in the \code{captures} data frame. Defaults to 330, the
approximate speed of sound in air.}
}
\description{
Simulates SCR capture histories and associated additional
information in the correct format for use with the function
\link{fit.acre}. If \code{fit} is provided then no other arguments
are required. Otherwise, at least \code{traps}, \code{mask}, and
\code{pars} are needed.
}
