% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_study.R
\name{sim_study}
\alias{sim_study}
\title{Title}
\usage{
sim_study(sim_name, n.rand = 1, fit = FALSE, seed = 810)
}
\arguments{
\item{sim_name}{}

\item{n.rand}{}

\item{fit}{}

\item{seed}{}
}
\description{
Title
}
