% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{ss_toa}
\alias{ss_toa}
\title{Data to showcase a "ss_toa" demo}
\format{
a list with all necessary input for the model:
\describe{
\item{detfn}{detechtion function: "ss" - signal strength}
\item{traps}{a data frame with the coordinates of the acoustic detectors}
\item{control.mask}{a list with the basic argument "buffer", which is the max detectable distance,
to create the detectable area from the coordinates of the detectors}
\item{ss.opts}{a list contains signal strength model related options. Here it contains "cutoff",
which indicates the threshold of a distance of 100\% detection.}
\item{capt}{a data frame of the capture history, contains columns as follows:\cr
\cr
"session" - the indices of the survey sessions;\cr
"ID" - the indices of the calls been detected;\cr
"occasion" - not been used, could be ignored;\cr
"trap" - the indices of the detectors;\cr
"ss" - the detected signal strength;\cr
"toa" - the time of arrival.\cr
}
\item{sv}{a list, contains the coefficient which been assigned a start value for modeling}
}
}
\source{
created from the simulation
}
\usage{
ss_toa
}
\description{
For the demonstration of the model with signal strength detection function, and signal strength
and time of arriaval as additional information.
}
\keyword{datasets}
