% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{stdEr.acreboot}
\alias{stdEr.acreboot}
\title{Title}
\usage{
\method{stdEr}{acreboot}(
  object,
  types = NULL,
  pars = NULL,
  new.covariates = NULL,
  from_boot = TRUE,
  show_fixed_par = TRUE,
  ...
)
}
\arguments{
\item{object}{an object generated by the bootstrap function "boot.acre()".}

\item{types}{a character vector, the same as "coef.acre()".}

\item{pars}{a character vector, the same as "coef.acre()".}

\item{new.covariates}{a data frame, the same as "coef.acre()".}

\item{from_boot}{a logical value, the same as "vcov.acreboot()".}

\item{show_fixed_par}{a logical value, the same as "vcov.acre()".}
}
\description{
Title
}
